test: organize workflow coverage suites
